Update Solution.java
i가 46341 이상이면 i * i가 int 범위 초과되므로 long 타입을 강제 적용해야한다.

+++ Codility/CountFactors/Solution.java
@@ -1,9 +1,9 @@
 class Solution {
     public int solution(int N) {
         int cnt = 0;
-        for(int i=1; i*i<=N; i++){
+        for(int i=1; (long)i*i<=N; i++){
             if((N%i)==0){
-                if(i*i==N){
+                if((long)i*i==N){
                     cnt++;
                 }else{
                     cnt+=2;
